+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Bug #118: Add missing ARIA attributes to `Dropdown`, `Menu`, and `Alert` (@WarLikeLaux)
 - Enh #155: Add `url` as alias for `link` in `Menu` and `Dropdown` items (@WarLikeLaux)
 - Enh #157: Improve psalm type annotations (@Tigrov)
+- Enh #143: Add conditional CSS class arrays to all `*Class()` methods in `Alert`, `Dropdown`, and `Menu` (@WarLikeLaux)
 
 ## 2.1.1 September 23, 2025